Listen to init socket before the creation of reexec process
We need to ensure that the init socket exists before the reexec process
tries to connect to it. Otherwise, there will be a race between the
parent and reexec process. If the parent process does not manage to
reach create and listen to the socket, then the reexec process will fail
to connect to it, since the socket does not even exist.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
PR: #15