Introduce support to boot up QEMU with intrd
Introduce initial support to boot QEMU guests with urunc. For the time
being, we do not have support for virtio-blk or shared fs, but only for
initrd. Therefore, the guest can get data only through initrd. Regarding
networking, we use virtio-net over the tap interface that we create with
urunc.

It is important to note that we spawn a generic QEMU VM and not a
microVM.

Overall, things we need to consider in the future:
- Support for virtio-blk
- Support for shared fs
- VM memory
- Support for QEMU's microvm

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>