Introduce support for booting Unikraft unikernels
Add support for Unikraft unikernels in urunc. Urunc expects that the
user specified cmdline should start with the name of the app that is
running in the unikernel. This name will be the first argument in
Unikraft's CLI options.

For the time being urunc can boot Unikraft on top of QEMU. Moreover,
we only support initrd. Virtio-blk or shared fs is not supported
yet.

Fixes: #6
Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
PR: #18