ci: Update shutdown flag
For the VM test workflow, the default behavior should be to
shutdown the VM after the test. If we debug stuff on a PR, we
should be able to change this to false, so that the VM stays online
and we can manually shut it down. Otherwise, it leaves stale VMs
on our infra.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
PR: https://github.com/nubificus/urunc/pull/31