ci: Remove generated ssh key after artifact creation
There are cases when the tests job fails. If we want to re-run the
specific test job, it will probably be placed in the same github runner.
The job id will remain the same, so it will try to overwrite the generated
ssh key.

This is probably a bug on the ci logic, as we should cleanup the
generated keys.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
PR: https://github.com/nubificus/urunc/pull/35