Create and apply seccomp filter for Hvt VMM
In an effort to increase security of unikernels running over Solo5-hvt,
we create and apply a seccomp filter. The filter allows only a small
subset of system calls that can be used from Solo5. However, this small
subset is still enough for its proper execution.

An important thing to note is that in case a non-permitted system call
is used, the whole process will receive a SIGSYS.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
PR: https://github.com/nubificus/urunc/pull/40