Enable and disable Seccomp filters for VMM, based on container's spec
Use the container's spec to determine if we should enable the seccomp
filters for Hvt, Qemu and Firecracker. If the container spec includes a
secomp field, then we should enbale seccomp filters. Otherwise, we
disable them.

The seccomp field in the container spec can be controlled with the
--security-opt command line option in nerdctl. By default, seccomp is
enabled, but using the unconfined keyword in the previous command line
option, seccomp will get disabled.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>
PR: https://github.com/nubificus/urunc/pull/40