Use /proc/self/mountinf to get container rootfs mount
Change the way we retrieve information regarding the container's rootfs
mount. THe most reliable way to retireve correct info regarding mounts
is reading the contents of /proc/self/mountinfo. Therefore, we parse
this file and check for the mount in the given path.

PR: https://github.com/nubificus/urunc/pull/48
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>