Use annotation and environment variable for devmapper use as a block
Introduce another annotation "com.urunc.unikernel.useDMBlock" and an
environment variable "USE_DEVMAPPER_AS_BLOCK" to indicate the usage
of the devmapper snapshot as a block device for the unikernel.

The annotation has more power than the environment variable. Therefore,
if the user specifies the annotation in the bima container the
environment variable is not taken into consideration. However, we might
want to change this in the future.

At last but not least, if none of the annotation or the environment
variable is set, then urunc does not take insto consideration any of
them. Therefore, urunc will continue with the rest of the checks. In
particular, it will check if the bima container does not include a
block image, if we are actually using devmapper, if the unikernel
supports block and the devmapper's FS. Based on all these checks urunc
will use devmapper as a block device.

PR: https://github.com/nubificus/urunc/pull/48
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>