Refactor the way we prepare the DM block device for the unikernel
Change the code that handles the check and unmounting of the
devicemapper block device to use it as a block device for the unikernel.
In particular, we check if the unikernel framework supports the
filesystem type of the device mapper snapshot, before we use it for the
unikernel rootfs.

PR: https://github.com/nubificus/urunc/pull/48
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>