Enhance unikernel interface with functions to check supporting features
Change the way we initialize a unikernel struct reference and add three
more functions in the unikernel interface. Instead of creating a new
unikernel struct directly, we break the creation and initialization of
the unikernel struct. THerfore, at first, we simply create a new
unikernel struct, based on the unikernel type and then we initialize it
with the Init function in the unikernel interface. The reason behind
this change is to allow the use of unikernel-specific functions (e.g.
SupportsBlock()) without initializing a new unikernel struct, so we can
use this function during the delete phase, as a check for devmapper
cleaning.

Moreover, we add three more functions in the unikernel interface:
- Init(): which initializes the unikernel struct based on the unikernel
arguments
- SupportsBlock(): which returns a bool value, based on the block
support of the respective unikernel.
- SupportsFS(); which takes as an argument a filesystem type and checks
if the unikernel supports that type.

PR: https://github.com/nubificus/urunc/pull/45
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>