Add annotations for block device and mountpoints
Add annotations for recognizing block devices inside the container. If
the user specifies a block device inside the bima container, urunc will
use this file as a block device when spawning the unikernel. Moreover,
another annotations is used to define the mountpoint of the block device
inside the unikernel.

It is important to note that the block device will be used only if the
unikernel supports Block devices.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>