Testing: Add test for bima containers that contain block files
THere are cases where the user might select to include in the bima
container the block file for the unikernel. In that case, we do not want
to use the devmapper, instead we want to directly use the storage file
inside the container.

This commit adds a test to check this new functionality in urunc.

PR: https://github.com/nubificus/urunc/pull/48
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>