refactor(testing): Make nerdctl tests run as subtests
Use an umbrella test for all nerdctl tests and transform the existing
tests into subtests.

In addition, refactor the code to use a test function driver which
prepares the environment for the test. In particular, it creates the
container and gives the control to the defined function test to perform
the specified test.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>