refactor(testing): Make ctr tests run as subtests
Use an umbrella test for all ctr tests and transform the existing tests
into subtests.

In addition, refactor the code to use a test function driver which
prepares the environment for the test and cleans up later. In
particular, it creates the container and gives the control to the
defined function test to perform the specified test.

At last, change the way we test with ctr by capturing and comparing the
output of the container, instead of just checking if the VMM process
started.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>