ref(testing): Make crictl tests run as subtests
Use an umbrella test for all crictl tests and transform the existing
tests into subtests.

In addition, refactor the code to use a test function driver which
prepares the enviornment for the test and cleans up later. In
particular, it creates the pod and the container and gives the control
to the defined function test to perform the specified test.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>