Update seccomp filter for Solo5-hvt VMM
In Ubuntu 22.04, there are new system calls that are used before the
execution of the solo5-hvt VMM. As a result, the VMM execution was
failing due to the fact that we were not allowing the execution of these
system calls.

This patch allows the execution of the extra required system calls.

Fixes https://github.com/nubificus/urunc/issues/49

PR: https://github.com/nubificus/urunc/pull/55
Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>