Refactor and enhance Makefile
Refactor Makefile to make it more modular and easier to update it.
Moreover, add additional building, linting, testing and help ragets.
Overall the changes are the following ones:
- Use variables instead of hardcoded commands
- Allow the definition of variables when invoking make
- Proper use of prerequisities for the targets
- Add target to build dynamically-linked urunc binaries
- Add targets to run specific end-to-end tests
- Add target for linting
- Remove use of sudo in rules. Instead the make command should be
invoked with sudo
- Add target for help message

Moreover, change the workflows to make use of make commands to run the
tests and not directly through go commands. THis will allow us to make
changes in the Makefile regarding the tests but we will not need to
change anything in the workflows.

PR: https://github.com/nubificus/urunc/pull/54
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>