Change the way we check for QueueProxy container
In the case of QueueProxy container we want to set some environment
variables and let runc to handle the container.However, in order to do
that we need to first identofy that we have a QueueProxy container. For
that purpose we had a function that was loading the container
specification and was checking a specific annotation.

We can replace this function and integrate the check in the usual
process of getting the information for the current container. In that
way we do not need to read the spec several times, but only just once
and depending on the annotations, we either continue or let runc to
handle it. Such an approach will reduce errors such the one we had with
k3s where we were using wrong path for the container bundle.

PR: https://github.com/nubificus/urunc/pull/57
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>