Use the correct bundle directory
In some cases (e.g. k3s) the bundle dir is not in /run/containerd, but
in some other location. When urunc gets invoked with the create command,
we receive the bundle directory through the command line options and we
store it in the state.json file. THerefore, we should either read the
command line argument or get the bundle dir from the state json file.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>