Set mask to 1 for rumprun network config
Rumprun seems to have issues when we set a mask different than one in
the case of k8s and k3s. For that reason, we change the mask to 1.
However, we might want to revisit this in the future.

Fixes https://github.com/nubificus/urunc/issues/50 for Rumprun

PR: https://github.com/nubificus/urunc/pull/57
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>