Change the way we check if we handle a typical container
Urunc can not handle normal containers and simply forwards their
handling to runc. However, in order to check if the current container is
a typical container urunc used to read the container's specification and
try to retrieve the urunc specific annotations. If these annotations
were not existing then urunc was letting runc to handle the container.

In the case the annotations existed, urunc was trying to reread the
annotations. Therefore, with this commit we read the annotations once
and decide if we will proceed or if we will let runc to handle the
container.

Such an approach reduces possible erros such as the issue where we were
not reading the correct bundle directory in the case of k3s.

Fixes: https://github.com/nubificus/urunc/issues/32
Fixes: https://github.com/nubificus/urunc/issues/50

PR: https://github.com/nubificus/urunc/pull/57
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>