Replace all hardcoded values in paths with constants
Using hardcoded values is a potential point of introducing bugs. In
addition, it makes it difficult to perform any changes. For thaese
reasons, this commit replaces all the hardcoded values of the paths with
constants.

Furthermore, the commit changes the way we retrieve the sandbox path in
case we need to join a namespace, since it was a hardcoded one.

PR: https://github.com/nubificus/urunc/pull/57
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>