Replace hardcoded annotations with constant values
This commit makes use of constants instead of hardcoded values to refer
to annotations. In that way we can easily change the name of an
annotation in the future, without too much effort.

It is important to mention, that since Golang does not allow us to use
constant strings as tags in structs, we must keep two references of the
annotations up to date and in sync. The two references are located in
the same file one after the other. The first is the annotation constants
and the second is the json tag in unikernelConfig struct.

PR: https://github.com/nubificus/urunc/pull/57
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>