hypervisors: Append console option to FC to enable boot on arm64
It seems that for the arm64 cases, Unikraft (<0.17.0) does not boot
on FC unless the console is initialized. If we don't specify
the cmdline option, the unikernel does not boot at all. If we disable
the console, we get an undefined reference.

https://github.com/unikraft/unikraft/blob/97523b55ac9e1327f7294da1feec8d4869da534d/plat/kvm/arm/setup.c#L166

```
unikraft/lib/nolibc/scanf.c:35:(.text+0x4a50): undefined reference to `ukplat_coutk'
```

PR: https://github.com/nubificus/urunc/pull/58
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>