fix(Makefile): Remove ignoring errors when running tests
The CI uses the make rules to execute the tests. However, in the
Makefile rules, we were ignoring the success/fail result of the test
commands. As a result, the CI could not report the fail correctly.

Thanks to Kostis Papazafeiropoulos for finding it.

PR: https://github.com/nubificus/urunc/pull/68
Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>