refactor(testing): Create an interface for the tools starting with ctr
Create an interface for all the tools that we use to test urunc. The
interface specifies a set of functions that all the tools need to
implement. This patch starts with ctr. In addition, introduce some
common implementations of these functions that can be used from the
tools.

Furthermore, check also for the k8s.io namespace of containerd for
possible stale files after removing the container.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>