fix(unikraft): Creation of kernel arguments
As mentioned in https://github.com/nubificus/urunc/issues/52, tt seems
that urunc does not handle correctly the single word command line
option. In particular, the first word of the command line option was
supposed to be used for the AppName field of Unikraft cli struct and the
rest of the command as the cli option. However, if the command line was
a single word, then the AppName was blank and neither the application
arguments were set according to the single word.

This patch fixes the above issue, by using the single word command line
option for the AppName field and the application arguments.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>