feat(unikraft): Bring initrd unikraft up-to-date with newer versions
Unikraft changed the way it sets the kernel arguments for initrd from
the v0.16.1. Therefore, similarly with the network options, we change
the format of the command line arguments to reflect the latest changes.

It is important to note, that currently we use a lot of hardcoded
values. We really need to change this in the future. However, a better
understanding of the new options is necessary. This can be possible
addressed when we add shared-fs.

PR: https://github.com/nubificus/urunc/pull/72
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>