refactor(testing): Use a temporary directory for each test
Crictl create two files for the container and pod config. We can use the
CWD to store these files. As such, we change directory in the beginning
of each test to a new temporary directory, which will get automatically
removed after the completion of the test.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>