refactor(testing): Change code file organization and improve error check
Cleanup the code in the end-to-end testing files and separate the test
skeleton from the test functions.

In addition pass the testing struct of each subtest in the driver
function for better handling of test cleanup and error logginf. At last,
improve the error checking and prining.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>