refactor(testing): Implement the toolTest interface for crictl
Add new functions in toolTest interface to accomondate the crictl tests.
Moreover, implement the interface for the crictl tool. Therefore, we can
now execute all the tests in the same way and we specify various
functions that can be used from the test functions to check if the
unikernel is running successfuly on top of urunc.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>