feat(spellcheck): Let the user disable/enable spellcheck with comments
Allow the user to be able to exclude either a line or a block of code
from the spellcheck linter.

The user can exclude:
- a line using <comment-char> spellchecker:disable-line
- a block of code setting in the beginning of the block
<comment-char> spellchecker:off
and in the end of the block
<comment-char> spellchecker:on

For example in Go we can disable spellchecking for the linter with
<code> // spellchecker:disable-line
or
// spellchecker:off
<code>
// spellchecker:on

It is important to note that only the "//" and "#" special characters
are supported to make a comment.

PR: https://github.com/nubificus/urunc/pull/74
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>