refactor(testing): Move struct and testMethod definition to skeleton
Move the definition of containerTestArgs and testMethod function from
the e2e_tests.go file to test_skeleton file. For some reason, the VS
Code Go plugin was not happy with the previous code organization.
Therefore, change it in order to satisfy it.

PR: https://github.com/nubificus/urunc/pull/73
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>