feat(mewz): Add tests for Mewz
Add end-to-end tests to check the functionality of Mewz unikernels
with urunc. In particular, we add tests for nerdctl, docker and crictl.
Unfortunately, there are not a lot of examples to use as test sna dhence
we only use a simple hello-world http server. We could also use a Hello
world example and capture output, however mewz shuts down in an uncommon
way and the qemu process returns 1 as exit status. As a result, the test
seems failed, even if it was running successfully.

PR: https://github.com/nubificus/urunc/pull/152
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>