fix(knative): Enable NAT and IP forwarding in static networking
In the case of Knative, urunc makes use of static networking
configuration for the unikernel/user-container. All the requests to the
user-container go though the queue-proxy container and static networking
has worked well for that case, by simply making queue proxy to use the
static IP of the unikernel, instead of the localhost.

However, there is also the case where the user-container might interract
directly with the outside world (e.g. making a new HTTP request). In
that scenarion, the static networking faces a limitation. The IP of the
unikernel is only known inside the network namespace of the container.
Therefore, as soon as the package leaves the network namespace and
reaches the destination, any reply will never get delivered to the
unikernel, since it is an unknown IP.

One solution for the above issue is using NAT for the outgoing traffic
of the unikernel's IP. THis commit implements this logic, using iptables
and enabling IP forwarding.In the future, we might want to revisit this
and either change the networking setup, or create the same logic without
the use of iptables.

Fixes: https://github.com/nubificus/urunc/issues/76

PR: https://github.com/nubificus/urunc/pull/81
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>