fix(testing): Handle warnings during container operations
The tools that testing uses to spawn and manage containers, such as
nerdctl docker and crictl, might print in stderr some warning messages.
An example is the case of nerdctl and the use of systemd cgroups. In
that scnearion, the testing code, which expects a specific output from
the commands, will not work properly. For instance, when creating a new
container and the warning is printed in stderr, the testing code treats
this output as a container id.

TO resolve this issue, this commit removes the sue of CombinedOutput and
separates stdout and stderr. If the execution of the command fails for
any reason, then we combine the stdout with stderr and return the
combined output.

PR: https://github.com/nubificus/urunc/pull/84
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>