fix(docs): Update yaml in kubernetes tutorial
The previous yaml was using an image that does not exist and it was also
specifying the node to deploy the container. Hence, as pointed out in
https://github.com/nubificus/urunc/issues/52#issuecomment-2561921038,
the deployment was not successful.

This commit updates the image and removes the nodeName field. Moreover,
a new nginx Rumprun image was created and addded in the list of existing
images.

PR: https://github.com/nubificus/urunc/pull/86
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>