feat(linux): Add initial support for Linux guests
Provide support for booting Linux VMs over Qemu and Firecracker to run
containers targetting Linux. The focus is more on single-application
containers which will be packed along with the Linux kernel.

Currently, this commit adds support for block and initrd support for
Linux guests.

Furthermore, it handles the Linux kernel boot parameters and uses the
container arguments as cli arguments for the init process. THe first
argument in the container configuration will be the init process.
However, Linux can not recognize multi-word arguments for init. To
address this, we follow a simple convention with urunit. All multi-word
cli arguments will get wrapped in single quotes. Therefore, the init
process hould be responsible to unwrap them and handle them correctly.
An implementation of such init is urunit.

PR: https://github.com/nubificus/urunc/pull/158
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>