feat(linux): Add tests for Linux
Add end-to-end tests to verify the proper functionality of urunc with
Linux guests. In particuler, we add several tests for all the tools,
testing both network connectivity and handling of environment and cli
arguments. In addition, we use all kind of support storage solutions for
Linux, such as block, devmapper and initrd.

At last, increate the timeout for the e2e testing to 4 minutes, since we
have added a lot of tests that exceed 3 minutes.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>