refactor: Execute VMM from inside the container's rootfs
Change the working directory before the execution of the VMM to be
inside the container's rootfs.

The main motivation behind this change is to enable the future support
of non-root execution of the VMMs. If we are outside the rootfs and not
a root user, we do not have permissions to read the contents of the
rootfs and hence neither the unikernel binary/initrd etc.

Furthemore, this change will allow us to better prepare for the support
of jailer execution of the VMM.

PR: https://github.com/nubificus/urunc/pull/92
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>