fix(devmapper): Correct rootfs path in devmapper preparation as a block
The rootfs of the container is not always under the bundle directory.
Instead, depending on which tool/service created the container, the
rootfs might be in any location in the system. Therefore, instead of
assuming the rootfs is always under the bundle, just use the rootfs
directory given by the spec.

Also, clean the bundle and rootfs paths before using them.

PR: https://github.com/nubificus/urunc/pull/92
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>