feat(namespaces): Proper handling of namespaces
This is a huge commit, but it is too much overhead to break it down in
smaller ones and it is not that easy either. So, let me explain what is
happening here.

First of all, we had a problem when entering a namespace due to the
threading model of Go. The setns system call (used by netns package --
github.com/vishvananda/netns) is thread-specific, meaning that only the
calling thread enters the namespace and not the other threads of the
process. Since Go is by design multi-threaded the above means that we
enter the network namespace with one thread but not with the
others. As a result, it was not always guaranteed that the creation of
the tap device or even the spawning of the unikernels will take place
from the thread in in the correct namespace. Long story short, we
were not entering correctly the network namespace of the pod.

An attempt to solve that would be to specify in the go runtime
to use a single OS thread and lock the "main" go thread in that single
thread. This could fix entering some of namespaces but not all. The user
and mount namespaces can not be entered by setns from multi-threaded
processes. Furthermore, there is literally no way to make Go run as
single-threaded. Hence, it is impossible to enter all namespaces with
pure Go code.

The solution was to use C and follow the same method as runc. Using C we
can now safely and correctly enter or create the respective namespaces
from the container's specification.

runc comes with nsenter, a C code that runs as an "init constructor" of
Go apps. In simpler words, the C piece of code always executes right
before the Go runtime even starts and hence we have a traditional single
threaded context. We can use the nsenter package of libcontainer by
simply importing it and by default it will always execute first.
However, if some environment variables are not set, it will simply exit
and return the control to the Go app.

Therefore, in order to integrate nsenter with urunc we had to change
the way we reexec. First of all, we have to create a socketpair for the
communication between urunc create and nsenter. This socketpair will be
used to send data to nsenter and receive from it. We pass the socketpair
from ExtraFiles in go's cmd and set up the respective environment
variables to trigger nsenter execution and inform it about the file
descriptors.

The communication between urunc create and nsenter takes place through
the aformentioned socketpair and using the netlink protocol. Following
similar approach with libcontainer, we format the namespace information
from the container;s spec in the respective netlink message and we send
it to nsenter. Then nsenter does its magic by creating and joining the
respective namespaces and sends back to urunc create the pids of the
processes that it used and lets the reexec program to execute.

To conlude, with this patch we add support for all namespaces, except
user. They require better handling in the cleanup path and we will
do it next iteration.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>