chore: Update linter and ci
Update the version of the linter in both the Makefile and the
flow. Furthermore, fix any new linting issues found from the newer
linter version.

ALso change the runner for the unit testing workflow. Instead of using the
gcc runner and install go, simply use the go runner.

Furthermore, disable building binaries for all architectures, since this
is not straightforward with CGO.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>