fix(ipc): Remove urunc's init socket
We used the init socket in order to let urunc create know that
the reexec process is up and running. However, with the use of nsenter,
we already know that the reexec process is up and running because of the
pids that the nsenter sends back to urunc create. Hence, we can remove
the init socket.

However, we might want to revisit this synchronization scheme, because
the reexec process must read the state file when the urunc create has
updated it with the correct pid (received from nsenter). Therefore, we
are currently waiting for the ack message from urunc create. However,
there might be a race here, because the urunc create tries to write in a
socket that has to be created from reexec. We currently solve this with
send and retry, but a more elegant solution could be explored.

PR: https://github.com/nubificus/urunc/pull/129
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>