fix(network): Cleanup the network in the correct namespace
As part of destroying a unikernel container, we need to clean up the
tap device and the rest networking stuff. However, it is important to be
in the container's namespace to perform the above steps. For that
reason, when urunc kill is called (which later does the network
cleanup), we instruct Go's runtime to use a single thread for our code
and we lock the main go routine to this thread.

In that way, it is safe to enter the network namespace with setns and
then do the cleanup.

TODO; We will need to revisit this part when we add support for user
namespaces. We can not enter a network namespace that belongs to a
different user namespace than the one that we are currently. Therefore,
we might have to use the nsenter once more.

PR: https://github.com/nubificus/urunc/pull/129
Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>
Approved-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>